requirements.txt
azure-identity
azure-mgmt-authorization
azure-mgmt-resource
msgraph-sdk